refactor: :truck: merge projects routes
