+++ apps/client/src/stores/project.spec.ts
@@ -5,10 +5,12 @@ import { useProjectStore } from './project.js'
 import { useUsersStore } from './users.js'
 
 const apiClientGet = vi.spyOn(apiClient.Projects, 'getProjects')
+const apiClientList = vi.spyOn(apiClient.Projects, 'listProjects')
 const apiClientPost = vi.spyOn(apiClient.Projects, 'createProject')
 const apiClientPut = vi.spyOn(apiClient.Projects, 'updateProject')
 const apiClientReplayHooks = vi.spyOn(apiClient.Projects, 'replayHooksForProject')
 const apiClientDelete = vi.spyOn(apiClient.Projects, 'archiveProject')
+const apiClientPatch = vi.spyOn(apiClient.Projects, 'patchProject')
 
 describe('Project Store', () => {
   beforeEach(() => {
@@ -139,5 +141,54 @@ describe('Project Store', () => {
     expect(apiClientDelete).toHaveBeenCalledTimes(1)
     expect(apiClientGet).toHaveBeenCalledTimes(1)
     expect(projectStore.projects).toEqual([])
+  })
+})
+
+describe('Project Admin Store', () => {
+  beforeEach(() => {
+    vi.resetAllMocks()
+    // creates a fresh pinia and make it active so it's automatically picked
+    // up by any useStore() call without having to pass it to it: `useStore(pinia)`
+    setActivePinia(createPinia())
+  })
+
+  it('Should get project list by api call', async () => {
+    const data = [
+      { id: 'id1', name: 'project1', status: 'archived', members: [{ userId: 'a', role: 'user', email: 'test@test.com' }] },
+      { id: 'id2', name: 'project2', status: 'created', members: [{ userId: 'a', role: 'user', email: 'test@test.com' }] },
+      { id: 'id3', name: 'project3', status: 'created', members: [{ userId: 'a', role: 'user', email: 'test@test.com' }] },
+    ]
+    apiClientList.mockReturnValueOnce(Promise.resolve({ status: 200, body: data }))
+    const projectStore = useProjectStore()
+
+    const res = await projectStore.getAllProjects({})
+
+    expect(res).toBe(data)
+    expect(apiClientList).toHaveBeenCalledTimes(1)
+  })
+
+  it('Should get active project list by api call', async () => {
+    const data = [
+      { id: 'id2', name: 'project2', status: 'created', members: [{ userId: 'a', role: 'user', email: 'test@test.com' }] },
+      { id: 'id3', name: 'project3', status: 'created', members: [{ userId: 'a', role: 'user', email: 'test@test.com' }] },
+    ]
+    apiClientList.mockReturnValueOnce(Promise.resolve({ status: 200, body: data }))
+    const projectStore = useProjectStore()
+
+    const res = await projectStore.getAllProjects({ statusNotIn: 'archived' })
+
+    expect(res).toStrictEqual(data)
+    expect(apiClientList).toHaveBeenCalledTimes(1)
+  })
+
+  it('Should lock or unlock a project', async () => {
+    const project = { id: 'id1', name: 'project1', status: 'archived', locked: true }
+    apiClientPatch.mockReturnValueOnce(Promise.resolve({ status: 200 }))
+    const projectStore = useProjectStore()
+
+    const res = await projectStore.handleProjectLocking(project.id, project.locked)
+
+    expect(res).toBe(undefined)
+    expect(apiClientPatch).toHaveBeenCalledTimes(1)
   })
 })
