+++ apps/server/src/resources/index-admin.ts
@@ -1,5 +1,4 @@
 import { type FastifyInstance } from 'fastify'
-import { projectAdminRouter } from './project/admin/router.js'
 import { quotaAdminRouter } from './quota/admin/router.js'
 import { stageAdminRouter } from './stage/admin/router.js'
 import { userAdminRouter } from './user/admin/router.js'
@@ -9,7 +8,6 @@ import { checkAdminGroup } from '@/utils/controller.js'
 
 export const apiRouterAdmin = () => async (app: FastifyInstance) => {
   app.addHook('preHandler', checkAdminGroup)
-  await app.register(serverInstance.plugin(projectAdminRouter()))
   await app.register(serverInstance.plugin(quotaAdminRouter()))
   await app.register(serverInstance.plugin(stageAdminRouter()))
   await app.register(serverInstance.plugin(userAdminRouter()))
