+++ apps/server/src/resources/project/controllers.spec.ts
@@ -1,12 +1,13 @@
-import { vi, describe, it, expect, beforeAll, afterEach, afterAll } from 'vitest'
+import { vi, describe, it, expect, beforeAll, afterEach, afterAll, beforeEach } from 'vitest'
 import prisma from '../../__mocks__/prisma.js'
 import { getRequestor, setRequestor } from '../../utils/mocks.js'
 import app from '../../app.js'
-import { createRandomDbSetup, getRandomProject, getRandomRole, getRandomUser } from '@cpn-console/test-utils'
+import { createRandomDbSetup, getRandomProject, getRandomRole, getRandomUser, repeatFn } from '@cpn-console/test-utils'
 import { faker } from '@faker-js/faker'
-import { descriptionMaxLength, projectIsLockedInfo } from '@cpn-console/shared'
+import { adminGroupPath, descriptionMaxLength, projectIsLockedInfo } from '@cpn-console/shared'
 import { getConnection, closeConnections } from '../../connect.js'
 import { rolesToMembers } from './business.js'
+import { json2csv } from 'json-2-csv'
 
 vi.mock('fastify-keycloak-adapter', (await import('../../utils/mocks.js')).mockSessionPlugin)
 vi.mock('@cpn-console/hooks', (await import('../../utils/mocks.js')).mockHooksPackage)
@@ -288,6 +289,96 @@ describe('Project routes', () => {
 
       expect(response.statusCode).toEqual(403)
       expect(JSON.parse(response.body).error).toEqual('Vous n’avez pas les permissions suffisantes dans le projet')
+    })
+  })
+})
+
+describe('Admin project routes', () => {
+  beforeAll(async () => {
+    await getConnection()
+  })
+
+  afterAll(async () => {
+    return closeConnections()
+  })
+
+  beforeEach(() => {
+    const requestor = { ...getRandomUser(), groups: [adminGroupPath] }
+    setRequestor(requestor)
+
+    vi.clearAllMocks()
+  })
+
+  afterEach(() => {
+    vi.clearAllMocks()
+  })
+
+  // GET
+  describe('generateProjectsDataController', () => {
+    it('Should retrieve all projects data for download', async () => {
+      const projects = repeatFn(2)(getRandomProject)
+
+      prisma.project.findMany.mockResolvedValue(projects)
+
+      const response = await app.inject()
+        .get('/api/v1/projects/data')
+        .end()
+
+      expect(response.statusCode).toEqual(200)
+      expect(response.body).toEqual(json2csv(projects, {
+        emptyFieldValue: '',
+      }))
+    })
+  })
+
+  // PATCH
+  describe('handleProjectLockingController', () => {
+    it('Should lock a project', async () => {
+      const project = getRandomProject()
+
+      // @ts-ignore
+      prisma.project.update.mockResolvedValue(project)
+
+      const response = await app.inject()
+        .patch(`/api/v1/projects/${project.id}`)
+        .body({ lock: true })
+        .end()
+
+      expect(response.statusCode).toEqual(200)
+    })
+
+    it('Should unlock a project if not failed', async () => {
+      const project = getRandomProject()
+      project.status = 'created'
+
+      prisma.environment.findMany.mockResolvedValue([])
+      prisma.repository.findMany.mockResolvedValue([])
+      // @ts-ignore
+      prisma.project.update.mockResolvedValue(project)
+
+      const response = await app.inject()
+        .patch(`/api/v1/projects/${project.id}`)
+        .body({ lock: false })
+        .end()
+
+      expect(response.statusCode).toEqual(200)
+    })
+
+    it('Should not unlock a project if failed', async () => {
+      const project = getRandomProject()
+      project.status = 'failed'
+
+      prisma.environment.findMany.mockResolvedValue([])
+      prisma.repository.findMany.mockResolvedValue([])
+      // @ts-ignore
+      prisma.project.update.mockResolvedValue(project)
+
+      const response = await app.inject()
+        .patch(`/api/v1/projects/${project.id}`)
+        .body({ lock: false })
+        .end()
+
+      expect(response.statusCode).toEqual(200)
     })
   })
 })
