+++ apps/server/src/resources/project/router.ts
@@ -7,9 +7,12 @@ import {
   getProjectSecrets,
   replayHooks,
   listProjects,
+  handleProjectLocking,
+  generateProjectsData,
 } from './business.js'
 import { projectContract } from '@cpn-console/shared'
 import { serverInstance } from '@/app.js'
+import { checkIsAdmin } from '@/utils/controller.js'
 
 export const projectRouter = () => serverInstance.router(projectContract, {
 
@@ -154,6 +157,38 @@ export const projectRouter = () => serverInstance.router(projectContract, {
     })
     return {
       status: 204,
+      body: null,
+    }
+  },
+  // Récupérer les données de tous les projets pour export
+  getProjectsData: async ({ request: req }) => {
+    checkIsAdmin(req.session.user)
+    const generatedProjectsData = await generateProjectsData()
+
+    addReqLogs({
+      req,
+      message: 'Données des projets rassemblées pour export',
+    })
+    return {
+      status: 200,
+      body: generatedProjectsData,
+    }
+  },
+
+  // (Dé)verrouiller un projet
+  patchProject: async ({ request: req, params, body: data }) => {
+    checkIsAdmin(req.session.user)
+    const projectId = params.projectId
+    const lock = data.lock
+
+    await handleProjectLocking(projectId, lock)
+
+    addReqLogs({
+      req,
+      message: `Projet ${lock ? 'verrouillé' : 'déverrouillé'} avec succès`,
+    })
+    return {
+      status: 200,
       body: null,
     }
   },
